get order client and server connection is implemented

+++ CommercialApplication/ApplicationLayer/Dtoes/Order/OrderItemDto.cs
@@ -5,6 +5,6 @@
         public int ProductId { get; set; }
         public int Amount { get; set; }
         public double DiscountBasic { get; set; }
-        public long ActionId { get; set; }
+        public int ActionId { get; set; }
     }
 }
